Implemented Non-Overlapping Triangulation
Modified add_vertex_to_convex_hull() in order to return
an array of the points that are visible to the new point
(including lower and upper tangents). These are the points
which have a line to the new point without crossing the convex hull.

Added code to generate_delaunay_triangulation() in order
to generate a non-overlapping triangulation from the visible
points returned by add_vertex_to_convex_hull(). There are no
indices so points are repeated.